requirements.txt
streamlit
Pillow
pandas
numpy
pytesseract
python-docx
fpdf
opencv-python
PyMuPDF
autocorrect
textblob
openpyxl
rapidfuzz
